Localized some global variables and changed readme
